Abstracts away AppRoutes and context types
