+++ src/components/App.tsx
@@ -1,13 +1,10 @@
 import React from "react";
-import Products from "./products/Products";
-import Navbar from "./Navbar/Navbar";
-import Cart from "./Cart/Cart";
-import Checkout from "./CheckoutForm/Checkout/Checkout";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { CircularProgress, ThemeProvider, useTheme } from "@mui/material";
+
+import { BrowserRouter as Router } from "react-router-dom";
+import { ThemeProvider, useTheme } from "@mui/material";
 
 import CartContextProvider from "@/context/CartContext";
-import { useCartContext } from "@/hooks/useCartContext";
+import AppRoutes from "./AppRoutes";
 
 const App = () => {
   const theme = useTheme();
@@ -20,53 +17,6 @@ const App = () => {
         </Router>
       </ThemeProvider>
     </CartContextProvider>
-  );
-};
-
-const AppRoutes = () => {
-  const {
-    cart,
-    order,
-    errorMessage,
-    handleCaptureCheckout,
-    handleUpdateCartQty,
-    handleRemoveFromCart,
-    handleEmptyCart,
-  } = useCartContext();
-
-  return (
-    <>
-      <Navbar totalItems={cart?.total_items} />
-      <Routes>
-        <Route path="/" element={<Products />} />
-        <Route
-          path="/cart"
-          element={
-            cart ? (
-              <Cart
-                cart={cart}
-                handleUpdateCartQty={handleUpdateCartQty}
-                handleRemoveFromCart={handleRemoveFromCart}
-                handleEmptyCart={handleEmptyCart}
-              />
-            ) : (
-              <CircularProgress />
-            )
-          }
-        />
-        <Route
-          path="/checkout"
-          element={
-            <Checkout
-              cart={cart}
-              order={order}
-              onCaptureCheckout={handleCaptureCheckout}
-              error={errorMessage}
-            />
-          }
-        />
-      </Routes>
-    </>
   );
 };
