+++ src/hooks/useCartContext.tsx
@@ -1,20 +1,6 @@
-import { Product } from "@/lib/types/products/types";
 import { createContext, useContext } from "react";
-import { Cart as CartInterface } from "@/lib/types/products/types";
-import { newOrder } from "@/lib/types/products/types";
 
-type CartContextType = {
-  products: Product[];
-  cart: CartInterface | null;
-  order: any;
-  errorMessage: string;
-  onAddToCart: (productId: string, quantity: number) => void;
-  handleUpdateCartQty: (productId: string, quantity: number) => void;
-  handleRemoveFromCart: (productId: string) => void;
-  handleEmptyCart: () => void;
-  handleCaptureCheckout: (checkoutTokenId: string, newOrder: newOrder) => void;
-  refreshCart: () => void;
-};
+import { CartContextType } from "@/lib/types/context/types";
 
 export const CartContext = createContext<CartContextType>({
   products: [],
